Leveled storage. Emulates cases when we have large and slow storage and we would like to promote data from it to faster (but smaller) storage. It emulates sparse files with higher than block size granularity
Leveled storage.
Emulates cases when we have large and slow storage and we would like to promote data from it to faster (but smaller) storage. It emulates sparse files with higher than block size granularity

Leveled storage. Emulates cases when we have large and slow storage and we would like to promote data from it to faster (but smaller) storage. It emulates sparse files with higher than block size granularity